+++ molly/thread.lua
@@ -118,7 +118,7 @@ end
 -- threads, which are scheduled by the runtime.
 local function scheduler()
     if current.scheduler ~= nil then
-        current.scheduler()
+        return current.scheduler()
     end
     return true
 end